web_backend/requirements-backend.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
PyYAML
orjson
pytest
httpx
